added commerceJs to app router and rename files js
